user-service/pyproject.toml
[tool.poetry]
name = "user_service"
version = "0.1.0"
description = ""
authors = ["Oleg Gorbatiuk <oleg36531@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
passlib = "^1.7.4"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
bcrypt = "^4.3.0"
black = "^25.1.0"
ruff = "^0.9.10"
mypy = "^1.15.0"
pytest = "^8.3.5"
types-passlib = "^1.7.7.20241221"
anyio = "^4.8.0"
pytest-asyncio = "^0.25.3"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.3"
twisted = "^24.11.0"
types-psycopg2 = "^2.9.21.20250121"
psycopg2-binary = "^2.9.10"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
